refactor(domain): clean up project structure and naming conventions
- Move enums to dedicated `Domain/Enums` folder
- Fix and standardize namespaces across the project
- Improve naming conventions for interfaces, entities, and files
- Refactor entity classes for better readability and consistency
