+++ Domain/Entities/Inventory.cs
