+++ Domain/Entities/PurchaseHistory.cs
