+++ Domain/Entities/Sales.cs
@@ -1,0 +1,8 @@
+﻿namespace Domain.entities;
+
+public class Sales
+{
+    public  int Id { get; set; }
+    public  int OrderId { get; set; }
+    public decimal TotalAmount { get; set; } 
+}
