+++ Domain/Enums/OrderStatus.cs
@@ -1,0 +1,8 @@
+namespace Domain.Enums;
+
+public enum OrderStatus
+{
+    Processing =1,
+    Shipped,
+    Delivered   
+}
