+++ Domain/Enums/UnitTypes.cs
@@ -1,0 +1,7 @@
+namespace Domain.Enums;
+
+public enum UnitTypes
+{
+    PerPeace=1,
+    PerKg
+}
