+++ Domain/Interfaces/IGenericRepository.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace OnlineStorAccess.DataAccessCls
+namespace Domain.Interfaces
 {
     public interface IGenericRepository <T> where T : class
     {
