+++ Domain/Interfaces/IpaymentMethod.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OnlineStorBuisness.Interfaces
+﻿namespace Domain.Interfaces
 {
     internal interface IpaymentMethod
     {
