+++ Domain/entities/person.cs
@@ -1,19 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Domain.entities;
 
-namespace Domain.entities
+public class Person
 {
-    public class Person
-    {
-        public int id {  get; set; }
-        public required string fName { get; set; }
-        public required string  lName { get; set; }
-         
-        public string? Phone { get; set; }
-        public string? Email { get; set; }
-
-    }
+    public int Id {  get; set; }
+    public required string FName { get; set; }
+    public required string  LName { get; set; }
+    public string? Phone { get; set; }
+    public string? Email { get; set; }
 }
